Applied façade design pattern to TopicUseCases
